added custom exceptions

+++ src/main/java/com/backend/ecommercebackend/authentication/service/impl/AuthenticationServiceImpl.java
@@ -15,10 +15,10 @@ import com.backend.ecommercebackend.model.user.User;
 import com.backend.ecommercebackend.repository.user.UserRepository;
 import com.backend.ecommercebackend.service.impl.EmailServiceImpl;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -27,6 +27,7 @@ import org.springframework.stereotype.Service;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -49,18 +50,21 @@ public class AuthenticationServiceImpl implements AuthenticationService {
         }
 
         if(request.getAcceptTerms().equals(Boolean.FALSE)){
-            throw new RuntimeException("accept terms can be choose");
+            throw new ApplicationException(Exceptions.TERMS_ACCEPTANCE_EXCEPTION);
         }
+
+        Optional<User> existingUser = repository.findByEmail(request.getEmail());
+        if (existingUser.isPresent()) {
+            throw new ApplicationException(Exceptions.USER_ALREADY_EXIST);
+        }
+
         User user = authMapper.RegisterDtoToEntity(request,passwordEncoder);
         user.setRole(Role.USER);
         user.setCreatedAt(LocalDateTime.now());
         user.setUpdatedAt(LocalDateTime.now());
-        try {
-            repository.save(user);
-        }
-        catch (DataIntegrityViolationException e) {
-            throw new ApplicationException(Exceptions.USER_ALREADY_EXIST);
-        }
+
+        repository.save(user);
+
         emailService.deleteStoredEmail(request.getEmail());
 
         String accessToken = jwtService.generateAccessToken(user.getEmail());
@@ -100,28 +104,27 @@ public class AuthenticationServiceImpl implements AuthenticationService {
         }
 
     @Override
-    public void refreshAuthToken(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    public AuthResponse refreshAuthToken(HttpServletRequest request) throws IOException {
         String userEmail;
         String refreshToken;
-        AuthResponse authResponse;
         String authHeader = request.getHeader("Authorization");
-        if(authHeader==null || !authHeader.startsWith("Bearer ")){
-            return ;
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            throw new ApplicationException(Exceptions.INVALID_TOKEN_EXCEPTION);
         }
         refreshToken = authHeader.substring(7);
-        userEmail=jwtService.extractUsername(refreshToken);
-        if(userEmail!=null){
-            var user =this.userDetailsService.loadUserByUsername(userEmail);
-            if(jwtService.validateToken(user,refreshToken)){
+        userEmail = jwtService.extractUsername(refreshToken);
+        if (userEmail != null) {
+            var user = this.userDetailsService.loadUserByUsername(userEmail);
+            if (jwtService.validateToken(user, refreshToken)) {
                 String accessToken = jwtService.generateAccessToken(userEmail);
-                authResponse=AuthResponse.builder()
+                return AuthResponse.builder()
                         .refreshToken(refreshToken)
                         .accessToken(accessToken).build();
-                new ObjectMapper().writeValue(response.getOutputStream(),authResponse);
+            } else {
+                throw new ApplicationException(Exceptions.USER_NOT_FOUND);
             }
-            else {
-              throw new ApplicationException(Exceptions.USER_NOT_FOUND);
-            }
+        } else {
+            throw new ApplicationException(Exceptions.INVALID_TOKEN_EXCEPTION);
         }
     }
 }
